*** HACK DETECTION ***
This program will detect if a hacked ROM is a simple binary hex-edit
of the source ROM. It does this by checking how many bytes are the
same between the two ROMs. If more than 50% of the common size is
the same, the ROM is assumed to be a binary edit.

The "common size" is the size of the smallest ROM. Padding bytes from
the source ROM are ignored, so it will likely be smaller than the
source ROM.

TODO:
- Fancy colorization on terminals on Windows and Linux.
- Unicode filenames on Windows.
- Round percentages down.