build-msvc.bat: Use the static MSVC runtime.
Something's wrong with my MinGW-w64 setup. The compiled executable
doesn't work on Windows XP or Windows 7, or even on Wine. It simply
starts and then exits.

I'll rebuild it using MSVC 2017 for now. Unfortunately, this means
the executable will be bigger, since it statically links the MSVC
runtime instead of using MSVCRT.DLL.